refactor: extract chat history domain helpers
